fix: centralize provider compatibility policy
